api/requirements.txt
# FastAPI and server
fastapi>=0.121.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.0
pydantic-settings>=2.7.0

# CORS and middleware
python-multipart>=0.0.12

# Environment variables
python-dotenv>=1.0.0

# Supabase client
supabase>=2.24.0

# JWT verification
python-jose[cryptography]>=3.3.0

# Database ORM and migrations
sqlmodel>=0.0.14
alembic>=1.13.0
psycopg2-binary>=2.9.9

# AWS S3
boto3>=1.35.0

# AI and NLP (will add later in Epic 3 & 4)
# langchain>=0.3.0
# langchain-openai>=0.2.0
# openai>=1.50.0
# spacy>=3.8.0

# HTTP client
httpx>=0.27.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Code quality
ruff>=0.7.0
black>=24.0.0
mypy>=1.13.0
